refactor WDL parsing into separate trait/package
