+++ core/src/main/resources/reference.conf
@@ -36,6 +36,12 @@ submissionmonitor {
   trackDetailedSubmissionMetrics = true
 }
 
+wdl-parsing {
+  cache-max-size = 2500
+  parser-thread-pool-size = 8
+  parser-thread-pool-timeout-seconds = 50
+}
+
 akka.http.host-connection-pool.max-open-requests = 16384
 akka.http.host-connection-pool.max-connections = 20
 akka.http.server.idle-timeout = 210 s
